refactor: move our non-headless archiver code into `general` package
