+++ internal/pkg/archiver/headless/headless.go
@@ -1,0 +1,7 @@
+package headless
+
+import "github.com/internetarchive/Zeno/internal/pkg/log"
+
+var logger = log.NewFieldedLogger(&log.Fields{
+	"component": "archiver.headless",
+})
